requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyJWT==2.8.1
bcrypt==4.1.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Werkzeug==3.0.1
